Set up wallaby and add a feature test
Followed the instructions in the wallaby readme, using chromdriver
because that seems to be the most common setup that works well for
people, based on looking around the elixir forum

I added a new mix alias for running tests in ci because you have to
ensure that js is built otherwise liveview stuff doesn't work. there may
be a day when we need to also add the css assets?

Finally, I skipped everything in the setup about running tests async and
ensuring database connections don't mess with one another. We may need
to do that stuff at some point.

Note that failing feature tests will add a screenshot artifact

closes #311