Add recently added to sort_by
This also does some refactoring of sort_by, so to add new things we just
update Solr.Constants.
https://hexdocs.pm/elixir/module-attributes.html#as-compile-time-constants
said that's pretty common. Maybe it should be elsewhere though? I dunno.

The macro business is a little complicated - basically we have to have
module attributes to use them in function guards, but you can't call
attributes from another module, so we have to compile-time define them
in places we want to use those constants. To do that we can do `use` and
`using`, effectively working as a module include from Ruby (hmmm, maybe
that's bad..), to make it available to the compiler.

We also have to
call Gettext's macro directly, because if I `use Gettext` in that `__using__` then it
complains for modules that already `use Gettext`. If we instead call
`gettext` with dynamic values, then GetText is unable to recognize those
strings for the mix tasks, and I like that workflow.

Work supporting #437