Some discussion about possible refactors
We spent some time thinking about how there's no out-of-broadway path to
converting from a Figgy resource to an Item for use in the display, and
thought that if there was one it might make it clearer where the
complexity in that pipeline is.

Here's some code and comments that share some of that discussion. Some
takeaways were that two big differences are that there's no way to
convert to a HydrationCacheEntry (because what we really generate is the
parameters to save one), and that there's
HydrationCacheEntry.to_solr_document and Item.from_solr, and having
consistency there might help.

Co-authored-by: Amin Zare <AminEdZare@users.noreply.github.com>
Co-authored-by: Anna Headley <hackartisan@users.noreply.github.com>
Co-authored-by: Eliot Jordan <eliotjordan@users.noreply.github.com>
Co-authored-by: Shaun Ellis <sdellis@users.noreply.github.com>