Add playwright to cache to speed up build.
This also adds GHA docker layer caching to speed up docker builds.

It's important that the cache has a different key if it's setting up
assets or not, otherwise the format task wins and builds a cache without
Playwright.

The build finishes in 3 minutes 14 seconds now, and when there's no new
things to add to the build the docker container builds in 24s (like on
merge to main.)