Make viewer take full space, don't capture mouse events (#539)
* WIP

* Increase size of viewer a tad, disable mouse events.

I want to do more, like disable mouse scroll but keep pan, but it looks
like to do that we either need to fix
https://github.com/openseadragon/openseadragon/issues/2151 or add a
placeholder canvas to Clover.

* Mix format.

* Cleanup.

* Display improvements.

This makes it cover up the footer and the search bar.

Right now the only problem is you can still tab to those hidden links.

* Hide/show elements on rendering viewer.

This way tabbing doesn't try to go through the things in the header and
footer.

* It's not a class, it's a selector.

* Pull viewer out of the item pane, hide it all.

* Add comments.