Use Claude to generate an initial mockup.
Pre-Prompt - Generate a style guide so it doesn't make up colors and
stuff:

You're in an Elixir Phoenix application. You're acting as a UX
signer. Our goal
today is to write a markdown document that represents the style guide
for this application. Look at app.css, BrowseLive, HomeLive, SearchLive, and
ItemLive, determine which styles, colors, and components are being used, and write a style
 guide in markdown which meets best practices. Do not change any code,
d do not mention colors or formatting which isn't in use in the application.

First Prompt:

> You're in a Phoenix LiveView application. You're acting as a UX
> designer. We're going to develop a collection page purely for design
> purposes called "South Asian Ephemera." It
> should have a collection page at /collections/sae. Read and follow the
> style guide at STYLE_GUIDE.md to guide decisions around colors,
> typography, and components. Don't be
> afraid to build new components in the new live view if necessary. The
> goal of the collection page is to get people interested in the
> collection contents so that they click
> through some real items. The page should have a way to browse all
> (/search?filter[project]=South+Asian+Ephemera), display recently updated
> items, and have an engaging hero
> portion that doesn't take up the entire screen height, but has two
> columns - one with a tagline, the title of the collection, and maybe
> some way to engage with the collection,
> and the second with some engaging examples of material in the
> collection laid out in an aesthetically pleasing and interesting way.
> When populating example items for this
> mockup, use metadata and images you can get from this URL:
> https://digital-collections.princeton.edu/search?filter[project]=South+Asian+Ephemera
> (fetch them.) There also needs
> to be a way to access a long form description of the collection which
> isn't very engaging, but must be somewhere. When it comes time to store
> that description, ask it for me and
> I'll paste it. Use that description to generate a more engaging
> tagline.

Second Prompt (asked for description, but I realized it'd be hard to
paste because there's formatting):

> Now I need to get the collection description from you so I can
create the LiveView. Let me ask for that first:

This is Claude:

> Could you please provide the long form description of the South Asian
> Ephemera collection? I'll use it to create an engaging tagline and
> include it as a non-prominent
> description section on the page.

Me:

> Instead of pasting it, get the information at
> https://dpul.princeton.edu/sae/about/south-asian-ephemera-collection.
> You'll need all the text there about the south asian ephemera
> collection. Be creative about how to include the information, but try
> not to change the text too much.

It didn't render which was the whole point, so let it thrash around
enough to make it render:

> The page doesn't display. Use mix compile or test at
> http://localhost:4000/collections/sae, then fix any problems

At this point I have something rendering. I could ask it to move things
around, but in my experience when you have a long conversation it tends
to make things better than worse. There's a starting point, I can
improve it now if I like where it started. If I don't, then I'd consider
this a failure.

My initial reaction is I like it pretty well. The colors are poorly set
up, but the mini cards with the little tilting effect was nice, and it
had some interesting tags. It did a little drop-down drawer for "Learn
More", which I think is interesting but could use a bunch of cleanup.

I haven't looked at the code yet, that'd be my next step.