Try a hydration optimization
It'd need a bunch of cleaning up, but the idea is that on a fresh
hydration we can safely ignore all resources we're not going to process
at query time up until whatever timestamp was at the end when we started
the hydration. Then we iterate over every one.

Might not be worth it, but I think it might put Hydration in the
minutes, which'd be cool to know about.