Fix browse card focus states.
The borders from focusing were being covered by `overflow-hidden`, which
was being added to handle the "why are images blurred" header. This
adjusts the flex boxes so that's not as required and makes it so that
the "Updated" header is also included in the link and focus border,
meaning the whole card gets focused on tab.

Closes #787